sua checkout

+++ src/main/java/com/example/demo/repository/AddressRepository.java
@@ -4,7 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.demo.model.Address;
 import java.util.List;
 
-public interface AddressRepository extends JpaRepository<Address, Integer> {
+public interface AddressRepository extends JpaRepository<Address, Long> {
     List<Address> findByAccountId(Integer accountId);
     Address findByAccountIdAndIsDefaultTrue(Integer accountId);
+
 }
